requirements.txt
#requests==2.18.*
#tox==3.2.*
#mock==2.0.*
PyHamcrest==1.9.*
#pytest==3.7.*
